pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deia"
version = "0.1.0"
description = "Development Evidence & Insights Automation - Learn from AI-assisted development sessions"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Dave E.", email = "contact@deiasolutions.com"}
]
maintainers = [
    {name = "Dave E.", email = "contact@deiasolutions.com"}
]
keywords = [
    "ai",
    "development",
    "claude",
    "cursor",
    "copilot",
    "knowledge-management",
    "session-logging",
    "human-ai-collaboration"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dateutil>=2.8",
    "requests>=2.28",
    "watchdog>=3.0",
    "scikit-learn>=1.3.0",
    "rapidfuzz>=3.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "pytest-xdist>=3.5",  # Parallel test execution
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "pytest-xdist>=3.5",
]

[project.urls]
Homepage = "https://github.com/deiasolutions/deia"
Documentation = "https://github.com/deiasolutions/deia/docs"
Repository = "https://github.com/deiasolutions/deia"
Issues = "https://github.com/deiasolutions/deia/issues"
Changelog = "https://github.com/deiasolutions/deia/CHANGELOG.md"

[project.scripts]
deia = "deia.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
deia = [
    "templates/*.md",
    "rules/*.json",
    "rules/*.md",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
